# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api eccomerce",
        "contact": {
            "email": "guilhermeieski@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pagamento/pix": {
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Processa pagamento pix",
                "description": "Processa um pedido e retorna uma mensagem de sucesso ou erro",
                "operationId": "processarPedidoPix",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idUsuario": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "114.413.859-08"
                                    },
                                    "carrinho": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "preco": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100
                                                },
                                                "quantidade": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "Produto 1"
                                                },
                                                "cor_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "tamanho_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "produtoId": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "vlrFrete": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15
                                    },
                                    "agencia": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "servico": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido processado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "img_qrcode": {
                                            "type": "string",
                                            "example": "https://example.com/qrcode.png"
                                        },
                                        "text_qrCode": {
                                            "type": "string",
                                            "example": "QR Code Text"
                                        },
                                        "message": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao processar pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao processar pedido"
                                        },
                                        "erro": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}